+++ gae/layers.py
@@ -66,10 +66,10 @@ class Layer(object):
 
 class GraphConvolution(Layer):
     """Basic graph convolution layer for undirected graph without edge labels."""
-    def __init__(self, input_dim, output_dim, adj, dropout=0., act=tf.nn.relu, **kwargs):
+    def __init__(self, input_dim, output_dim, adj, dropout=0., act=tf.nn.relu, dtype=tf.float32, **kwargs):
         super(GraphConvolution, self).__init__(**kwargs)
         with tf.variable_scope(self.name + '_vars'):
-            self.vars['weights'] = weight_variable_glorot(input_dim, output_dim, name="weights")
+            self.vars['weights'] = weight_variable_glorot(input_dim, output_dim, dtype=dtype, name="weights")
         self.dropout = dropout
         self.adj = adj
         self.act = act
@@ -87,10 +87,10 @@ class GraphConvolution(Layer):
 
 class GraphConvolutionSparse(Layer):
     """Graph convolution layer for sparse inputs."""
-    def __init__(self, input_dim, output_dim, adj, features_nonzero, dropout=0., act=tf.nn.relu, **kwargs):
+    def __init__(self, input_dim, output_dim, adj, features_nonzero, dropout=0., act=tf.nn.relu, , dtype=tf.float32, **kwargs):
         super(GraphConvolutionSparse, self).__init__(**kwargs)
         with tf.variable_scope(self.name + '_vars'):
-            self.vars['weights'] = weight_variable_glorot(input_dim, output_dim, name="weights")
+            self.vars['weights'] = weight_variable_glorot(input_dim, output_dim, dtype=dtype, name="weights")
         self.dropout = dropout
         self.adj = adj
         self.act = act
